Arch setup script
Updates to clean up some paths to reference my standard dotfiles path,
swap to https instead of ssh for autojump for consistency,
and mostly, add a script for configuring a new arch distro.

This should also make a decent start for a new universal setup script
as it somewhat attempts to universal install methods except for where
pacman is the best way.